Convert the thousand group to an integral...
... before processing it. This avoids incorrectly entering the if block
when the group is not zero, but less than one (only decimals).
Note that the division by 1000 is already performed on the value as
long, as I suspect that it is less costly than floating-point division.

+++ src/main/java/io/github/moonstroke/spelledoutnumbers/impl/UsEnglishNumberSpeller.java
@@ -196,11 +196,11 @@ public class UsEnglishNumberSpeller implements NumberSpeller {
 				}
 				/* Clip the current thousands group on both sides to avoid inaccurate rounding
 				 * in big numbers */
-				double thisGroup = (doubleValue % 1e6) / 1000.;
+				long thisGroup = (long) (doubleValue % 1e6) / 1000;
 				if (thisGroup > 0) {
 					String rankName = getThousandsRankName(i);
 					words.add(rankName);
-					spellOutThousandGroup((long) thisGroup, words);
+					spellOutThousandGroup(thisGroup, words);
 				}
 				doubleValue /= 1000;
 			}
